+++ src/components/App.js
@@ -2,16 +2,16 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Link } from 'react-router-dom';
 import { MuiThemeProvider } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
+import Icon from '@material-ui/core/Icon';
 import MenuItem from '@material-ui/core/MenuItem';
 import Toolbar from '@material-ui/core/Toolbar';
 
-import ErrorBoundary from '../atoms/ErrorBoundary';
+import ErrorBoundary from 'atoms/ErrorBoundary';
 
 import Home from './Other/Home.js';
 import TalentProfile from './Profile/TalentProfile.container';
 import Contacts from './Contacts/Contacts.container';
 import theme from 'styles/theme';
-import { Icon } from 'semantic-ui-react';
 
 import GeneratedResumePage from 'pages/GeneratedResumePage/GeneratedResumePage.container';
 import CreateResumePage from 'pages/CreateResumePage';
@@ -21,11 +21,12 @@ const App = () => (
     <MuiThemeProvider theme={theme}>
       <Router>
         <React.Fragment>
+
           <AppBar position="static">
             <Toolbar>
               <Link to="/">
                 <MenuItem>
-                  <Icon name="home" />
+                  <Icon>home</Icon>
                 </MenuItem>
               </Link>
               {/* TODO change all urls to lowercase */}
@@ -40,6 +41,7 @@ const App = () => (
               </Link>
             </Toolbar>
           </AppBar>
+
           <Switch>
             {/*<Route path="/" component={AuthPage} />*/}
             {/*<Route exact path='/' component={AuthPage} />*/}
@@ -52,6 +54,7 @@ const App = () => (
             <Route exact path="/Contacts/:contactId/create-resume" component={CreateResumePage} />
             <Route exact path="/Contacts/:contactId/create-resume/:resumeId" component={CreateResumePage} />
           </Switch>
+
         </React.Fragment>
       </Router>
     </MuiThemeProvider>
