--- src/components/Profile/profile.css
@@ -1,80 +1,0 @@
-.item {
-    padding-left: 20px;
-    float: left;
-    min-width: 150px ;
-    margin: 10px 20px;
-    /*border-bottom: 1px solid grey;*/
-}
-.column{
-    display: flex;
-    flex-direction: column;
-}
-
-.infoBlock{
-    padding-left: 20px;
-    float: left;
-    min-width: 150px ;
-    margin: 10px 20px;
-    border-bottom: 1px solid grey;
-}
-.profileHeader{
-    background-color: "#f9f1de";
-    height: "230px";
-    /*display: flex;
-    flex-direction: row;*/
-    overflow: hidden;
-    min-width: 200px;
-
-}
-
-.rightCol{
-    display: flex;
-    flex-direction: row;
-}
-.allSkills{
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-.float_left{
-    float: left;
-    clear:left;
-}
-.float_right{
-    float: right; 
-    clear:right;
-}
-.skill{
-    width: 300px;
-    overflow: hidden;
-}
-
-
-
-
-.btn-circle.btn-xl {
-    width: 70px;
-    height: 70px;
-    padding: 10px 16px;
-    border-radius: 35px;
-    font-size: 24px;
-    line-height: 1.33;
-}
-
-.btn-circle {
-    width: 30px;
-    height: 30px;
-    padding: 6px 0px;
-    border-radius: 15px;
-    text-align: center;
-    font-size: 12px;
-    line-height: 1.42857;
-}
-
-/* Stupid hacks to make SemanticUI play nice with MaterialUI for the moment */
-.ui.form textarea {
-    border: none;
-    padding: 0;
-    box-sizing: border-box;
-    resize: none;
-}
