+++ src/constants.js
@@ -5,7 +5,7 @@ const getApiUrl = (environment) => {
     case 'staging':
       return 'http://dev-api.baltimorecorps.org';
     default:
-      return 'http://127.0.0.1:5000';
+      return 'http://localhost:5000';
   }
 };
