+++ src/modules/Achievements/AchievementsList/AchievementsListItem.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemText from '@material-ui/core/ListItemText';
+
+const AchievementsListItem = ({text}) => {
+  return (
+    <ListItem>
+      <ListItemText primary={text} />
+    </ListItem>
+  );
+};
+
+export default AchievementsListItem;
