+++ src/modules/Achievements/AchievementsList/index.js
@@ -1,0 +1,3 @@
+import AchievementsList from './AchievementsList';
+
+export default AchievementsList;
