+++ src/modules/Experiences/AddOrEditExperienceForm/index.js
@@ -1,0 +1,3 @@
+import AddOrEditExperienceForm from './AddOrEditExperienceForm';
+
+export default AddOrEditExperienceForm;
