+++ src/modules/Experiences/ExperiencesList/index.js
@@ -1,0 +1,3 @@
+import ExperiencesList from './ExperiencesList.container';
+
+export default ExperiencesList;
