+++ src/modules/Resumes/ResumesList/index.js
@@ -1,0 +1,3 @@
+import ResumesList from './ResumesList.container';
+
+export default ResumesList;
