+++ src/modules/Tags/AddOrEditSkillForm/index.js
@@ -1,0 +1,3 @@
+import AddOrEditSkillForm from './AddOrEditSkillForm.container';
+
+export default AddOrEditSkillForm;
