+++ src/modules/Tags/SkillsList/index.js
@@ -1,0 +1,3 @@
+import SkillsList from './SkillsList.container';
+
+export default SkillsList;
